+++ src/IframeProvider/IframeProvider.ts
@@ -1,19 +1,22 @@
-import type { SignableMessage } from '@terradharitri/sdk-core/out';
-import type { Transaction } from '@terradharitri/sdk-core/out/transaction';
-import {
-  WindowProviderResponseEnums,
-  ReplyWithPostMessagePayloadType
-} from '@terradharitri/sdk-dapp-utils/out';
+import { Message } from '@terradharitri/sdk-core/out';
+import { Transaction } from '@terradharitri/sdk-core/out/transaction';
 import {
   CrossWindowProvider,
-  ICrossWindowWalletAccount
+  IProviderAccount
   // !!! IMPORTANT !!! It is necessary to import explicitly from the file because the module exports (can export) some classes
   // that are using window API and will break the build process on the SSR environments (e.g. PopupConsent)
 } from '@terradharitri/sdk-web-wallet-cross-window-provider/out/CrossWindowProvider/CrossWindowProvider';
 import {
+  WindowProviderRequestEnums,
+  WindowProviderResponseEnums
+} from '@terradharitri/sdk-web-wallet-cross-window-provider/out/enums';
+import {
   ErrCouldNotLogin,
-  ErrProviderNotInitialized
+  ErrCouldNotSignTransactions,
+  ErrProviderNotInitialized,
+  ErrTransactionCancelled
 } from '@terradharitri/sdk-web-wallet-cross-window-provider/out/errors';
+import { ReplyWithPostMessagePayloadType } from '@terradharitri/sdk-web-wallet-cross-window-provider/out/types';
 import { IframeLoginTypes } from '../constants';
 import { IframeManager } from '../IframeManager/IframeManager';
 
@@ -46,8 +49,6 @@ export class IframeProvider extends CrossWindowProvider {
 
   public override async init(): Promise<boolean> {
     const initialized = await super.init();
-    await this.windowManager.setWalletWindow();
-
     return initialized;
   }
 
@@ -65,7 +66,7 @@ export class IframeProvider extends CrossWindowProvider {
     options: {
       token?: string;
     } = {}
-  ): Promise<ICrossWindowWalletAccount> {
+  ): Promise<IProviderAccount> {
     await this.windowManager.setWalletWindow();
     const account = await super.login(options);
 
@@ -74,6 +75,8 @@ export class IframeProvider extends CrossWindowProvider {
       this.windowManager.walletWindow = null;
       throw new ErrCouldNotLogin();
     }
+
+    this.windowManager.closeWalletWindow();
 
     return account;
   }
@@ -104,28 +107,64 @@ export class IframeProvider extends CrossWindowProvider {
     transaction: Transaction
   ): Promise<Transaction> {
     await this.windowManager.setWalletWindow();
-    return super.signTransaction(transaction);
+    const data = await super.signTransaction(transaction);
+    this.windowManager.closeWalletWindow();
+    return data;
   }
 
   public override async signTransactions(
     transactions: Transaction[]
   ): Promise<Transaction[]> {
     await this.windowManager.setWalletWindow();
-    return super.signTransactions(transactions);
+    this.ensureConnected();
+
+    const {
+      type,
+      payload: { data: signedPlainTransactions, error }
+    } = await this.windowManager.postMessage({
+      type: WindowProviderRequestEnums.signTransactionsRequest,
+      payload: transactions.map((tx) => tx.toPlainObject())
+    });
+
+    if (error || !signedPlainTransactions) {
+      this.windowManager.closeWalletWindow();
+      throw new ErrCouldNotSignTransactions();
+    }
+
+    if (type === WindowProviderResponseEnums.cancelResponse) {
+      this.windowManager.closeWalletWindow();
+      throw new ErrTransactionCancelled();
+    }
+
+    const hasTransactions = signedPlainTransactions?.length > 0;
+
+    if (!hasTransactions) {
+      throw new ErrCouldNotSignTransactions();
+    }
+
+    const data = signedPlainTransactions.map((tx) =>
+      Transaction.fromPlainObject(tx)
+    );
+
+    this.windowManager.closeWalletWindow();
+    return data;
   }
 
   public override async guardTransactions(
     transactions: Transaction[]
   ): Promise<Transaction[]> {
     await this.windowManager.setWalletWindow();
-    return super.guardTransactions(transactions);
+    const data = await super.guardTransactions(transactions);
+    this.windowManager.closeWalletWindow();
+    return data;
   }
 
-  public override async signMessage(
-    message: SignableMessage
-  ): Promise<SignableMessage> {
+  public override async signMessage(messageToSign: Message): Promise<Message> {
     await this.windowManager.setWalletWindow();
-    return super.signMessage(message);
+    messageToSign.signer = this.loginType;
+    const data = await super.signMessage(messageToSign);
+    this.windowManager.closeWalletWindow();
+    return data;
   }
 
   public override async openPopupConsent(): Promise<boolean> {